* commit 'b224b0293e59bd3ea2eb4c4fc6a3a5bfaf227d27':
  Fixed GCC warnings
  Fixed a few programming errors caught by GCC

+++ src/dsbexec/config_parser.cpp
@@ -8,6 +8,7 @@
 
 #include "boost/foreach.hpp"
 #include "boost/lexical_cast.hpp"
+#include "boost/numeric/conversion/cast.hpp"
 #include "boost/property_tree/info_parser.hpp"
 #include "boost/property_tree/ptree.hpp"
 
@@ -152,7 +153,7 @@ void ParseSystemConfig(
         const auto& slaveType = slaveTypes.find(slaveTypeName)->second;
         
         // Assign a slave ID
-        const auto slaveID = slaves.size() + 1;
+        const auto slaveID = boost::numeric_cast<uint16_t>(slaves.size() + 1);
         SlaveInstance slaveInstance = { slaveID, &slaveType };
         slaves[slaveName] = slaveInstance;
